Mon/Financial_Data_and_Performance_handling/backend/requirements.txt
langchain
langchain-openai
langchain-pinecone
langchain-redis
redis
python-dotenv
celery
fastapi
uvicorn
python-multipart
uuid
